feat: makes parent component smart
